requirements.txt
amadeus==8.0.0
asgiref==3.7.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
decorator==5.1.1
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-extra-fields==3.7.0
filelock==3.13.1
filetype==1.2.0
fonttools==4.49.0
future==0.18.3
geocoder==1.38.1
gunicorn==20.0.4
idna==3.4
isodate==0.6.0
itypes==1.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mysqlclient==2.2.3
numpy==1.26.4
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pycodestyle==2.5.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2019.3
qrcode==7.4.2
ratelim==0.1.6
requests==2.31.0
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
virtualenv==20.25.0
whitenoise==5.0.1
